move hamburger button to right
