+++ frontend/src/App.tsx
@@ -5,12 +5,13 @@ import Home from "./pages/Home"
 import "./App.scss"
 import Header from "./Header/Header"
 import Hamburger from "./Hamburger/Hamburger"
+import Menu from "./Menu/Menu"
 
 function App() {
   return (
     <BrowserRouter>
-      <Hamburger />
       <Header />
+      <Hamburger />
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/explore" element={<Explore />} />
