+++ frontend/src/Hamburger/Hamburger.tsx
@@ -2,7 +2,6 @@ import React, { useState } from "react"
 import "./Hamburger.scss" // styling for the hmauburger button icon
 
 import Menu from "../Menu/Menu"
-import { menuItems } from "../Menu/MenuItems"
 
 const Hamburger: React.FC = () => {
   const [isOpen, setIsOpen] = useState(false)
@@ -11,7 +10,7 @@ const Hamburger: React.FC = () => {
     setIsOpen(!isOpen);
   }
   return (
-    <div className="show-hamburger">
+    <div className="-showhamburger">
       <button onClick={handleClick} className="hamburger" data-testid="toggle-button">
         <div className="line" />
         <div className="line" />
@@ -19,8 +18,8 @@ const Hamburger: React.FC = () => {
       </button>
 
       {isOpen && (
-        <div className="menu" data-testid="menu" onClick={handleClick}>
-          <Menu items={menuItems} />
+        <div className="menu" data-testid="menu">
+          <Menu />
         </div>
       )}
     </div>
