+++ frontend/src/Header/Header.tsx
@@ -1,14 +1,9 @@
 import "./Header.scss"
 import { useNavigate } from "react-router-dom"
+import MenuItems from "../Menu/MenuItems"
 
 const Header: React.FC = () => {
   const logo = require("../images/Whale3logo.png")
-
-  const navigate = useNavigate()
-
-  const handleClick = () => {
-    navigate("/login")
-  }
 
   return (
     <div className="show-header">
@@ -17,23 +12,7 @@ const Header: React.FC = () => {
           <a className="navbar-brand" href="/">
             <img src={logo} alt="Whale Whale Whale logo" width="200" />
           </a>
-
-          <ul className="navbar-nav mb-2 mb-lg-0 d-flex w-100 justify-content-around">
-            <li className="nav-item">
-              <a className="nav-link active" aria-current="page" href="/">
-                Home
-              </a>
-            </li>
-            <li className="nav-item">
-              <a className="nav-link" href="/explore">
-                Explore
-              </a>
-            </li>
-          </ul>
-
-          <button className="btn btn-outline-success px-4" style={{ width: "200px" }} onClick={handleClick}>
-            Log in / Sign up
-          </button>
+          <MenuItems />
         </div>
       </nav>
     </div>
